camera bug fixed for both x and y directions, I think. Hard to tell on x direction.

+++ main.py
@@ -272,7 +272,7 @@ class Camera:
 			if (abs(xdiff) < mincammove):
 				newposx += xdiff
 			else:
-				xdeltamove = int(xdiff * camerasmoothmovespeedx)
+				xdeltamove = xdiff * camerasmoothmovespeedx
 				if (abs(xdeltamove) < mincammove):
 					xdeltamove = mincammove * sign(xdeltamove)
